backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
bcrypt
pyjwt
PyGithub
google-generativeai
apscheduler
httpx
slowapi
